general block entity support for all versions (didn't check 1.18 actually)
